+++ lib/anycable/grpc.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module AnyCable
+  module GRPC
+  end
+end
+
+require "anycable/grpc/server"
